Make group optional in object references
Allow omitting the group specification in object references, and document
default values for group and resource.

Before this commit, the group could be the empty string but could not be
omitted (that is, left unspecified in json or yaml definitions).

* api/v1alpha1/gateway_types.go (Listener): Document the default values for
group and resource for the listener extension object reference (the
defaults are core and configmaps, respectively).
(ListenerTLS): Document the default values for group and resource for
certificate object references (the defaults are core and secrets,
respectively).
(LocalObjectReference): Make group optional.
* api/v1alpha1/gatewayclass_types.go (GatewayClassSpec): Document the
default values for group and resource for the parameters object
reference (the defaults are core and configmaps, respectively).
* api/v1alpha1/httproute_types.go (HTTPRouteHost): Document the default
values for group and resource for the extension object reference (the
defaults are core and configmaps, respectively).
(HTTPRouteMatch): Document the default values for group and resource for
the extension object reference (the defaults are core and configmaps,
respectively).
(HTTPRouteFilter): Document the default values for group and resource for
the extension object reference (the defaults are core and configmaps,
respectively).
(HTTPRouteAction): Document the default values for group and resource for
the target object reference (for which the defaults are core and services,
respectively) and for the extension object reference (for which the
defaults are core and configmaps, respectively).
* api/v1alpha1/generated.proto:
* config/crd/bases/networking.x-k8s.io_gatewayclasses.yaml:
* config/crd/bases/networking.x-k8s.io_gateways.yaml:
* config/crd/bases/networking.x-k8s.io_httproutes.yaml: Regenerate.