Update Gateway status and conditions.
The structure for exposing Listener status was broken in #193 because
the Listener Name field was removed. This means that status can no longer
be associated to a Listener by matching the name.

However, since Listeners are basically metadata wrapped around ports,
we can expose granular Listener status by publishing a ListenerStatus
object for each port that is bound on the Gateway. This lets us model
port conflicts in a reasonable way as well as scope other errors more
tightly than the entire gateway.

In practice, the scoping of Listener conditions is not as granular as
previously, since most applications will all be exposed on port 443. In
this case, it may not be easy to distinguish which Listener's routes
were not ready.

Since IP addresses are a property of the Gateway, they should also be
a property of the GatewayStatus. Note that the status field contains
the bound addresses, not the requested addresses.

Condition types have been redistributed between the GatewayConditionType
and ListenerConditionType in accordance with the responsibility for the
error that each condition signifies.

Signed-off-by: James Peach <jpeach@vmware.com>