Hoist HTTP match types into type-specific structs.
Update the route match struct to hoist the path and header match
specifications into their own structs. This improves API clarity and
consistency by using the same matching syntax for paths and headers.

Update the matching defaults at various levels so that there is always a
default prefix path match on "/". This reduces the amount of YAML needed
for simple cases and removes the validation burden from controllers.

Signed-off-by: James Peach <jpeach@vmware.com>