HTTPRoute, TCPRoute: Add gateway status information
* apis/v1alpha1/httproute_types.go (HTTPRouteStatus): Change the slice of
GatewayObjectReference to a slice of RouteGatewayStatus.
(RouteGatewayStatus): New type.  Define a reference to a Gateway object and
a list of status conditions for the route with respect to that Gateway.
(RouteConditionType): New type for route status condition type.
(ConditionRouteAdmitted): New route status condition type.  A route status
condition with this type indicates whether the route has been admitted or
rejected by a Gateway.
(RouteCondition): New type for route status conditions.
* apis/v1alpha1/tcproute_types.go (TCPRouteStatus): Add a Gateways field
with type slice of RouteGatewayStatus.
* apis/v1alpha1/generated.pb.go:
* apis/v1alpha1/generated.proto:
* apis/v1alpha1/zz_generated.deepcopy.go:
* config/crd/bases/networking.x-k8s.io_httproutes.yaml:
* config/crd/bases/networking.x-k8s.io_tcproutes.yaml:
* docs-src/spec.md:
* docs/spec/index.html: Regenerate.