Standardize on the metav1.Condition type.
Replace all the custom condition types with `metav1.Condition`. This
reduces the amount of type duplication, at the cost of weaker typing for
specifying types in the context of the status of particular resources.
However, the weaker typing has the benefit that it gives implementations
more leeway to add their own condition types to expose more operational
information to users.

Signed-off-by: James Peach <jpeach@vmware.com>