Fix codegen script to work with Go modules.
To make the codegen script work wil Go modules, we first need to add
the generator module dependencies by referencing them in tools.go.
There doesn't seem to be a c0.19.0 release, so I bumped everything
up to v0.19.2.

The generator always writes to the $GOPATH, so faking one up in the
script is necessary for developers who don't check out the repository
in the traditional Go source hierarchy.

Finally, switch to "go run" so as to not pollute the $GOBIN directory,
format the command arguments consistently and apply shellcheck
suggestions.

Signed-off-by: James Peach <jpeach@vmware.com>