Linking from GitHub Pages to Netlify, removing generated docs from repo
Since Netlify can build our docs for us, there should be no need for us
to retain the generated files in this repo. This removes a lot of the
custom logic for building docs since it will no longer be necessary. It
also adds back search functionality since we no longer need to worry
about conflicts.