Add default value '/' for HTTPPathMatch
`HTTPPathMatch` has a default `prefix match on the "/" path`.
For example, if `match` was not provided, the default value will be filled.

(This configuration works.)
```
spec:
  rules:
```

However, once we added `type` such as `Prefix` or `Exact`, the `value`
field must be filled by ourselves.

(This configuration does not work.)
```
spec:
  rules:
    matches:
    - path:
        type: Prefix
```

This patch adds the default `/` value for `path.value`.