tag latest releases in the staging repo
The latest tag in the staging repository gives an easy way to test the
latest and greatest image of the admission-webhook-server.
This is primarily intended for maintainers of this repository and for
implementers of this API who might want to target their CI systems
against this tag.

The make target has been renamed from `release` to `release-staging` for
self-documenting code purpose and makes it obvious that the release is
for staging environment.