Removing scope fields from object references
Our object references use kind instead of resource which requires
implementations to have a predefined kind->resource mapping. That means
there can never be ambiguity within an implementation as far as whether
a reference is to a cluster-scoped or namespace-scoped resource.

This also removes the ability for ParentRef to refer to cluster-scoped
resources. Although we can add that in the future with a docs update,
there do not appear to be any use cases for that today.