fix: kustomize pathing to stable CRDs
A recent commit changed the directory name for the stable CRDs from
"v1alpha2" to "stable". The result is that CI I had running that was
deploying the CRDs via kustomize is now failing. This patch changes
the resource list of the kustomize configuration to point to the new
location of the CRDs.