Fix ReferencePolicy docs about setting empty value in `spec.to.name`
Current docs explains about `spec.to.name` as:

```
Name is the name of the referent. When unspecified or empty, this policy
refers to all resources of the specified Group and Kind in the local
namespace.
```

However, the ReferencePolicy with empty value:

```shell
kubectl apply -f - <<EOF
apiVersion: gateway.networking.k8s.io/v1alpha2
kind: ReferencePolicy
metadata:
  name: allow-gateways-to-ref-secrets
  namespace: default
spec:
  from:
  - group: gateway.networking.k8s.io
    kind: Gateway
    namespace: istio-system
  to:
  - group: ""
    kind: Secret
    name: ""
EOF
```

get the following error.

`The ReferencePolicy "allow-gateways-to-ref-secrets" is invalid: spec.to.name: Invalid value: "": spec.to.name in body should be at least 1 chars long`

It should be fine to support only "unspecified" case. So, this patch removes "empty" from docs.