docs: display beta before alpha in API spec
We display the alpha types in the API spec before the beta types. So
readers see all of our types - including experimental and deprecated -
before seeing the package with the more limited set of "standard"
types.

Reverse this by manually iterating the packages list so that the
beta package is displayed before the alpha package.