chore: replace "element in set" helper function
In our validation code we had added our own utility for
determining if a SectionName existed in a set of SectionNames,
however apimachinery recently added generics to pkg/util/sets
which can handle this for us. This patch updates the dependencies
so that we have access to those new utilities and replaces our
custom utility with the upstream sets utilities.