Use distinct names for test resources for test isolation
Two conformance tests are currently using the same name for their Gateway resource. This leads to contention when one test is actively deleting its Gateway and the other test is actively modifying it. The second test then attempts to retrieve the Gateway; however, the deletion has completed by that time, causing the second test to fail.