Promote Gateway for Mesh (GAMMA) to Standard (#2873)
This PR raises Gateway API for Mesh to "Standard" channel.

What is included:
* GEP-1426 (Routes binding to service) and GEP-1686 (conformance tests
  for mesh)
* Updated various docs to declare stable.

Followup needed:
* Blog post announcement

What is *not* included:
* Changes to any tests; we already have existing conformance for mesh,
  and the `feature` is already called `MeshCoreFeatures` so I don't
think we need to change anything there. In the future, we may add
experimental features to Mesh, at which point we can make a new set of
features like `MeshExperimentalFeatures` or similar.
* Changes to any Go/CRD APIs beyond docs. The Gateway for Mesh work did not
  introduce any API changes. Instead it just introduced new ways to use
the existing APIs with corresponding documentation changes to remove the
'Experimental' signifier.